Fix: Внесены исправления по комментариям

+++ src/main/java/ru/karyeragame/paymentsystem/bankaccount/repository/BankAccountRepository.java
@@ -5,8 +5,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import ru.karyeragame.paymentsystem.bankaccount.model.BankAccount;
 
+import java.math.BigDecimal;
+
 public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
 
     @Query("SELECT ba.balance FROM BankAccount ba WHERE ba.owner.id = :ownerId AND ba.type = :type")
-    Float findBalanceByOwnerIdAndType(@Param("ownerId") Long ownerId, @Param("type") String type);
+    BigDecimal findBalanceByOwnerIdAndType(@Param("ownerId") Long ownerId, @Param("type") String type);
 }
