+++ src/main/java/ru/karyeragame/paymentsystem/bankaccount/service/BankAccountService.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 
 public interface BankAccountService {
 
-    Float getBalanceByUserId(Long userId, BankAccountType type);
+    BigDecimal getBalanceByUserId(Long userId, BankAccountType type);
 
     BankAccount getBankAccountEntity(Long bankAccountId);
